UniLumos/UniLumos/requirements.txt
# Core dependencies for UniLumos inference
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
numpy==1.26.3
decord==0.6.0
av==13.0.0
opencv-python==4.11.0.86
Pillow==11.1.0
imageio==2.37.0
tqdm==4.67.1
einops==0.8.1
transformers==4.53.0
ftfy==6.3.1
regex==2024.11.6
diffusers
mmengine==0.10.7
pandas==2.2.3
pyarrow==19.0.1
requests==2.32.3
